refactor(ci/agentic-reviewer): extract shared env.py module
Consolidate the duplicated required_env/bool_env helper functions
from 7 modules into a single odh_ci_agent.env module.
