+++ ci/agentic-reviewer/src/odh_ci_agent/env.py
@@ -1,0 +1,16 @@
+"""Shared environment-variable helpers."""
+
+from __future__ import annotations
+
+import os
+
+
+def required_env(name: str) -> str:
+    value = os.environ.get(name, "").strip()
+    if value:
+        return value
+    raise SystemExit(f"Missing required environment variable: {name}")
+
+
+def bool_env(name: str) -> bool:
+    return os.environ.get(name, "").strip().lower() in {"1", "true", "yes", "on"}
